Preemption bug lead to a long busy waiting loop in heavy system load
At very high system load, the 2nd byte of a word read transfer
could lead to a 200ms+ busy wait if the preemption happened between
the data register read and the status register reset.
In this case, the interrupt flag was cleared as the transfer was
finished, leading to a never happening interrupt flag condition.